requirement.txt
annotated-types==0.5.0
anyio==3.7.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
fastapi==0.101.1
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
idna==3.4
packaging==23.1
proto-plus==1.22.3
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic_core==2.4.0
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
